Correction3 - Solution to Task 8

+++ 0x00-hello_world/101-quote.c
@@ -7,6 +7,6 @@
 int main(void)
 {
 write(2, "and that piece of art is useful\" - Dora Korpar, 2015-10-19\n",
-(unsigned long)sizeof("and that piece of art is useful\" - Dora Korpar, 2015-10-19"));
+sizeof("and that piece of art is useful\" - Dora Korpar, 2015-10-19"));
 return (1);
 }
